01092019 demo_page renamed to not_logged_in / minor design updates
