+++ client/src/not_logged_in/components/team.js
@@ -79,7 +79,7 @@ class Team extends React.Component {
                 <div className="row">
                     <div className="col-md-12 text-center">
                         <div className="section-title">
-                            <h2>Our Team</h2>
+                            <h2>Coders Nearby</h2>
                             <img src="assets/images/title-line.png" alt="title-line" className="img-fluid"/>
                         </div>
                     </div>
